Consolidate per-resource _http.py into single inkbox._http
The four resource packages (identities, mail, phone, vault) each had a
near-identical copy of HttpTransport. Merge them into one shared
inkbox/_http.py and update all imports. Matches the existing TS SDK
layout (single src/_http.ts shared across resources).
